Init Optimisation, Bugfixes, Readability
Moved more functionality over from the main to the Critter class (such as spawning randomly and bouncing off walls).

Textures are no longer loaded and unloaded when a critter is destroyed or respawned. Instead, there are only 2 textures loaded, which the critters all reference using pointers.

Upon being spawned in for the first time, critters have their textures set and their half-widths/heights adjusted accordingly.

Critter::Draw() now uses the critter's half-width/height to center the image. These variables are set when the texture is set.

m_radius is replaced with m_hWidth for all calculations involving radius. This is a temporary solution.